+++ README.md
@@ -72,7 +72,14 @@ character sets (UTF-8) and [terminal capabilities][] are taken advantage of.
   - [x] PostgreSQL for data
   - [x] Asynchronous ORM ([GINO][])
   - [x] User model
-  - [x] Message bases
+  - [ ] Message bases
+    - [x] List messages
+    - [x] Post messages
+    - [x] Reply to messages
+    - [x] Tag system
+    - [ ] Search messages
+    - [ ] Filter by tag(s)
+    - [ ] Private messages
   - [ ] Simple pickle table for miscellaneous data storage?
 
 </details>
