+++ userland/models/__init__.py
@@ -1,5 +1,13 @@
 """Default userland models"""
 
+from .message import Message
+from .message.message_tags import MessageTags
+from .message.tag import MessageTag
 from .oneliner import Oneliner
 
-__all__ = ("Oneliner",)
+__all__ = (
+    "Message",
+    "MessageTag",
+    "MessageTags",
+    "Oneliner",
+)
