+++ xthulu/models/__init__.py
@@ -1,13 +1,5 @@
 """Database ORM models"""
 
-from .message import Message
-from .message.message_tags import MessageTags
-from .message.tag import MessageTag
 from .user import User
 
-__all__ = (
-    "User",
-    "Message",
-    "MessageTag",
-    "MessageTags",
-)
+__all__ = ("User",)
